pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai"
version = "0.1.0"
description = "OpenAI agentic AI app"
requires-python = ">=3.8"
dependencies = [
    "openai-agents",
    "requests",
    "scikit-learn",
    "ddtrace"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]